[SYCL] Prevent use of fp64 and fp16 when unsupported in more tests (intel/llvm-test-suite#1353)
A number of tests currently use double and sycl::half despite them not
being supported by the used device. Although it does not necessarily
fail to run on those devices, they are not considered supported by the
SYCL specification and should not be used in testing on those devices.
This commit prevent parts of a selection of tests from running on
devices that do not support the types used.

+++ SYCL/Regression/local-arg-align.cpp
@@ -24,15 +24,15 @@ int main(int argc, char *argv[]) {
   buffer<size_t, 1> res(2);
 
   q.submit([&](sycl::handler &h) {
-     // Use two local buffers, one with an int and one with a double4
+     // Use two local buffers, one with an int and one with a float4
      local_accessor<cl_int, 1> a(1, h);
-     local_accessor<double4, 1> b(1, h);
+     local_accessor<float4, 1> b(1, h);
 
      auto ares = res.get_access<access::mode::read_write>(h);
 
      // Manually capture kernel arguments to ensure an order with the int
-     // argument first and the double4 argument second. If the two arguments are
-     // simply laid out consecutively, the double4 argument will not be
+     // argument first and the float4 argument second. If the two arguments are
+     // simply laid out consecutively, the float4 argument will not be
      // correctly aligned.
      h.parallel_for(1, [a, b, ares](sycl::id<1> i) {
        // Get the addresses of the two local buffers
@@ -52,10 +52,10 @@ int main(int argc, char *argv[]) {
     ret = -1;
   }
 
-  if (hres[1] % sizeof(double4) != 0) {
+  if (hres[1] % sizeof(float4) != 0) {
     std::cout
         << "Error: incorrect alignment for argument b, required alignment: "
-        << sizeof(double4) << ", address: " << (void *)hres[1] << std::endl;
+        << sizeof(float4) << ", address: " << (void *)hres[1] << std::endl;
     ret = -1;
   }
